针对非公开测试点修改了一些小bug 如show finance无交易记录 以及无输出时输出换行的问题

+++ src/BookSystem.h
@@ -71,7 +71,7 @@ private:
 
   std::fstream fin_log_file;
 
-  std::vector<double> income, outcome;
+  std::vector<long double> income, outcome;
 
   int fin_log_cnt;
